api: we're going with the generation approach to feed page saving
In brief: if we save an additional generation int every time we assignPages (basically previous generation + 1), we can tell repo impls to pick the lower generation number if there's two pages with no next link. This allows us to drop the requirement that saving pages must be atomic over everything. Instead, it only has to be atomic per list element which is easier to implement (if not vastly so for most DBs) and easier to explain since it's not a special case any more.