Feature/snapshot integrationtest (#15)
* Create SnapshotAcceptanceTest

* api: some javadoc

* api: some tests

* api: implement and test final Body methods

* api: HttpHeader impl and test

* api: impl, docs, test

* api: page docs

* api: allow comparing Body.fromBytes and Body.fromUtf8 return values

* api: Page test

* api: don't keep snapshot id in SnapshotPageHeader, but we do have to pass it to SnapshotPageRepository

* api: snapshot docs and test

* api: we're going with the generation approach to feed page saving

In brief: if we save an additional generation int every time we assignPages (basically previous generation + 1), we can tell repo impls to pick the lower generation number if there's two pages with no next link. This allows us to drop the requirement that saving pages must be atomic over everything. Instead, it only has to be atomic per list element which is easier to implement (if not vastly so for most DBs) and easier to explain since it's not a special case any more.

* api: feed docs

* api: The Forgotten Tests

* Fix code style

* Disable short variable name rule

* Fixed compilation failure

* SnapshotConsumerIntegrationTest

---------

Co-authored-by: Felix Krull <felix.krull@neuland-bfi.de>